forgot to commit last night, and fiddled around and added a bunch broken code this morning. Might have to reroll.
adding:
-avg arnona per sqmt per locale
-use value to estimate the arnona for each listing
-sort dataframes before plotting
-sort dataframes before listing results

+++ settings_manager.py
@@ -96,11 +96,15 @@ def delete_settings():
             continue
         break
 
-    if os.path.exists(filename):
-        os.remove(filename)
-        print("Settings deleted successfully.\n")
+    x = input("Are you sure you want to delete " + filename + " Setting? (y/n)")
+    if x == 'y':
+        if os.path.exists(filename):
+            os.remove(filename)
+            print("Settings deleted successfully.\n")
+        else:
+            print("Settings file does not exist.\n")
     else:
-        print("Settings file does not exist.\n")
+        return
 
     with open(filename, "w+") as fh:
         fh.close()
